+++ src/handlers/user_commands.ts
@@ -1,17 +1,13 @@
 import vscode from 'vscode';
 import { Auth0URLs, NOTIFICATION, WebPaths } from "../constants";
-import { createRedirectUri, generateAuthUrl, generateWebUrl } from "../utils/url_utils";
+import { createRedirectUri, generateAuthUrl, generateLogoutUrl, generateWebUrl } from "../utils/url_utils";
 import { markUsersInactive } from '../utils/auth_utils';
 import { UserState } from '../utils/user_utils';
 
 
 export const authHandler = (skipAskConnect=false) => {
 	(global as any).skipAskConnect = skipAskConnect;
-	const redirectURI = createRedirectUri(Auth0URLs.LOGIN_CALLBACK_PATH);
-	const additionalParams = {
-		"login-callback": redirectURI
-	};
-	const authUrl = generateWebUrl(WebPaths.AUTH, additionalParams);
+	const authUrl = generateAuthUrl();
 	const userState = new UserState();
 	userState.setWaitingForLogin();
 	vscode.window.showInformationMessage(NOTIFICATION.WAITING_FOR_LOGIN_CONFIRMATION);
@@ -20,16 +16,8 @@ export const authHandler = (skipAskConnect=false) => {
 
 
 export const logoutHandler = () => {
-    const logoutUrl = generateAuthUrl(Auth0URLs.LOGOUT);
-    vscode.env.openExternal(vscode.Uri.parse(logoutUrl));
-    markUsersInactive();
-	if ((global as any).socketConnection) {
-		(global as any).socketConnection.close();
-		(global as any).socketConnection = null;
-	}
-	if ((global as any).websocketClient) {
-		(global as any).websocketClient = null;
-	}
+    const logoutUrl = generateLogoutUrl();
+    vscode.env.openExternal(vscode.Uri.parse(decodeURIComponent(logoutUrl)));
     return logoutUrl;
 };